Update traineeView and add modelerror to system validator

+++ Controllers/StaffController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Ajax.Utilities;
 using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.EntityFramework;
+using Microsoft.Owin.Security;
 using System;
 using System.Collections.Generic;
 using System.Data.Entity;
@@ -119,9 +120,12 @@ namespace WebApplication1.Controllers
                 .Where(t => t.TraineeId == id)
                 .ToList();
             _context.Users.Remove(traineeaccount);
-            foreach (var trainee in traineeCourse)
+            if (traineeCourse != null)
             {
-                _context.TraineeCourses.Remove(trainee);
+                foreach (var trainee in traineeCourse)
+                {
+                    _context.TraineeCourses.Remove(trainee);
+                }
             }
             _context.SaveChanges();
 
@@ -174,10 +178,10 @@ namespace WebApplication1.Controllers
             edit_infor.UserName = Function.SplitUserComponents(changeTraineeDetaill.Trainee.UserName);
             edit_infor.Main_programming_lang = changeTraineeDetaill.Trainee.Main_programming_lang;
 
-            _userManager.RemovePassword(edit_infor.Id);
-            _userManager.AddPassword(edit_infor.Id, changeTraineeDetaill.Password);
+            _userManager.RemovePassword(changeTraineeDetaill.Id);
+            _userManager.AddPassword(changeTraineeDetaill.Id, changeTraineeDetaill.Password);
 
-            _userManager.UpdateAsync(edit_infor);
+            _userManager.Update(edit_infor);
             _context.SaveChanges();
 
             return RedirectToAction("TraineeAccountView");
@@ -456,6 +460,64 @@ namespace WebApplication1.Controllers
 
             return RedirectToAction("DetailCourse", "Staff", new { @id = traineeCourses.CourseId});
         }
+        #region Helpers
+        // Used for XSRF protection when adding external logins
+        private const string XsrfKey = "XsrfId";
+
+        private IAuthenticationManager AuthenticationManager
+        {
+            get
+            {
+                return HttpContext.GetOwinContext().Authentication;
+            }
+        }
+
+        private void AddErrors(IdentityResult result)
+        {
+            foreach (var error in result.Errors)
+            {
+                ModelState.AddModelError("", error);
+            }
+        }
+
+        private ActionResult RedirectToLocal(string returnUrl)
+        {
+            if (Url.IsLocalUrl(returnUrl))
+            {
+                return Redirect(returnUrl);
+            }
+            return RedirectToAction("Index", "Home");
+        }
+
+        internal class ChallengeResult : HttpUnauthorizedResult
+        {
+            public ChallengeResult(string provider, string redirectUri)
+                : this(provider, redirectUri, null)
+            {
+            }
+
+            public ChallengeResult(string provider, string redirectUri, string userId)
+            {
+                LoginProvider = provider;
+                RedirectUri = redirectUri;
+                UserId = userId;
+            }
+
+            public string LoginProvider { get; set; }
+            public string RedirectUri { get; set; }
+            public string UserId { get; set; }
+
+            public override void ExecuteResult(ControllerContext context)
+            {
+                var properties = new AuthenticationProperties { RedirectUri = RedirectUri };
+                if (UserId != null)
+                {
+                    properties.Dictionary[XsrfKey] = UserId;
+                }
+                context.HttpContext.GetOwinContext().Authentication.Challenge(properties, LoginProvider);
+            }
+        }
+        #endregion
     }
 
     class Function
